fix(avo-1724): copy flowplayer from react-comp, fix cuep. , active item
https://meemoo.atlassian.net/browse/AVO-1724
