+++ src/components/FlowPlayer/FlowPlayer.declared.d.ts
@@ -1,0 +1,11 @@
+declare module '@flowplayer/player/plugins/chromecast';
+declare module '@flowplayer/player/plugins/airplay';
+declare module '@flowplayer/player/plugins/subtitles';
+declare module '@flowplayer/player/plugins/speed';
+declare module '@flowplayer/player/plugins/audio';
+declare module '@flowplayer/player/plugins/hls';
+declare module '@flowplayer/player/plugins/cuepoints';
+declare module '@flowplayer/player/plugins/keyboard';
+declare module '@flowplayer/player/plugins/playlist';
+declare module '@flowplayer/player/plugins/google-analytics';
+declare module 'react-scrollbars-custom';
