+++ src/components/FlowPlayer/FlowPlayer.helpers.ts
@@ -1,3 +1,5 @@
+import { GoogleAnalyticsEvent } from './FlowPlayer.types';
+
 export function setPlayingVideoSeekTime(seekTime: number): void {
 	const playingVideo: HTMLVideoElement | null = document.querySelector(
 		'.c-video-player .is-playing video'
@@ -17,3 +19,11 @@ export function getPlayingVideoSeekTime(): number | null {
 		return null;
 	}
 }
+
+export const convertGAEventsArrayToObject = (googleAnalyticsEvents: GoogleAnalyticsEvent[]) => {
+	return googleAnalyticsEvents.reduce((acc: any, curr: GoogleAnalyticsEvent) => {
+		acc[curr] = curr;
+
+		return acc;
+	}, {});
+};
