+++ src/components/FlowPlayer/FlowPlayer.mock.tsx
@@ -1,12 +1,12 @@
-import { FlowPlayerPropsSchema, FlowplayerSourceList } from './FlowPlayer.types';
+import { FlowPlayerProps, FlowplayerSourceList } from './FlowPlayer.types';
 
-export const MOCK_FLOW_PLAYER_PROPS: FlowPlayerPropsSchema = {
+export const MOCK_FLOW_PLAYER_PROPS: FlowPlayerProps = {
 	src: 'http://commondatastorage.googleapis.com/gtv-videos-bucket/sample/ElephantsDream.mp4',
 	title: 'Title',
 	autoplay: false,
 };
 
-export const MOCK_FLOW_PLAYER_PROPS_FULL: FlowPlayerPropsSchema = {
+export const MOCK_FLOW_PLAYER_PROPS_FULL: FlowPlayerProps = {
 	...MOCK_FLOW_PLAYER_PROPS,
 	poster: 'https://via.placeholder.com/1920x1080',
 	logo: 'https://via.placeholder.com/100x100',
