+++ src/components/FlowPlayer/index.ts
@@ -1,0 +1,3 @@
+export { FlowPlayer } from './FlowPlayer';
+export * from './FlowPlayer.types';
+export * from './FlowPlayer.helpers';
