+++ src/react-app-env.d.ts
@@ -2,13 +2,3 @@
 
 declare module 'braft-editor';
 declare module 'braft-extensions/dist/table';
-declare module '@flowplayer/player/plugins/chromecast';
-declare module '@flowplayer/player/plugins/airplay';
-declare module '@flowplayer/player/plugins/subtitles';
-declare module '@flowplayer/player/plugins/speed';
-declare module '@flowplayer/player/plugins/hls';
-declare module '@flowplayer/player/plugins/cuepoints';
-declare module '@flowplayer/player/plugins/keyboard';
-declare module '@flowplayer/player/plugins/playlist';
-declare module '@flowplayer/player/plugins/google-analytics';
-declare module 'react-scrollbars-custom';
